// backend/src/types/generated/public/Users.ts
export type UsersId = string & { __brand: 'public.users' };

/** Represents the table public.users */
export default interface Users {
  id: UsersId;

  email: string | null;

  username: string | null;

  display_name: string | null;

  first_name: string | null;

  last_name: string | null;

  phone: string | null;

  photo_url: string | null;

  role: string;

  email_verified: boolean | null;

  phone_verified: boolean | null;

  is_active: boolean | null;

  yandex_id: string | null;

  yandex_email: string | null;

  personal_data: Record<string, unknown> | null;

  preferences: Record<string, unknown> | null;

  linked_accounts: Record<string, unknown> | null;

  created_at: string | null;

  updated_at: string | null;

  telegram_id: string | null;

  telegram_username: string | null;

  password_hash: string | null;

  skills: string[] | null;

  last_password_change: string | null;

  force_password_change: boolean | null;

  accept_calls: boolean | null;

  two_factor_enabled: boolean | null;

  two_factor_method: string | null;

  google_id: string | null;

  apple_id: string | null;

  vk_id: string | null;

  sber_id: string | null;

  mts_id: string | null;

  hired_date: string | null;

  last_seen_at: string | null;
}